Fixes files and syllabus scanning
the conversion to using a background worker
had a bunch of cleanup to the scanning code.
It incorrectly dealt with files and syllabus.
Thats all fixed now.

Files were trying to use the auth header in the api
requests, but they come with a temporary auth.

And syllabus api was incorrectly assuming it had
to loop of the results, which it doesnt.