Add javascript testing and gitlab-ci config
Added the jest framework and a single render
test as an example for adding more tests.

In order for the test to work, it was necessary
to refactor how the initialData was being populated
in the App components.

Also added the gitlab-ci.yml file to make tests run
on each commit.