Fix focus-management issues
Every time a pill button was pressed it would
remove itself, but the focus was not being set
on the previous element. This change sets the
focus on the previous element.

Also fixed a few bugs where alerts were not
rendering correctly and the UDoItModal was
erroring if the preview html was not present.

references https://cidilabs.atlassian.net/browse/UDOIT3-131